+++ dio/lib/src/adapters/browser_adapter.dart
@@ -51,10 +51,12 @@ class BrowserHttpClientAdapter implements HttpClientAdapter {
 
     final connectTimeout = options.connectTimeout;
     final receiveTimeout = options.receiveTimeout;
+    int xhrTimeout = 0;
     if (connectTimeout != null &&
         receiveTimeout != null &&
         receiveTimeout > Duration.zero) {
-      xhr.timeout = (connectTimeout + receiveTimeout).inMilliseconds;
+      xhrTimeout = (connectTimeout + receiveTimeout).inMilliseconds;
+      xhr.timeout = xhrTimeout;
     }
 
     final completer = Completer<ResponseBody>();
@@ -176,6 +178,21 @@ class BrowserHttpClientAdapter implements HttpClientAdapter {
         ),
         StackTrace.current,
       );
+    });
+
+    xhr.onTimeout.first.then((_) {
+      if (connectTimeoutTimer != null) {
+        connectTimeoutTimer?.cancel();
+      }
+      if (!completer.isCompleted) {
+        completer.completeError(
+          DioException.receiveTimeout(
+            timeout: Duration(milliseconds: xhrTimeout),
+            requestOptions: options,
+          ),
+          StackTrace.current,
+        );
+      }
     });
 
     cancelFuture?.then((_) {